@@ -91,7 +91,7 @@
     "ejs": "^2.6.1",
     "ejs-webpack-loader": "^2.2.2",
     "enzyme": "^3.9.0",
-    "enzyme-adapter-react-15.4": "^1.3.1",
+    "enzyme-adapter-react-15": "^1.4.0",
     "eslint": "^4.19.1",
     "eslint-config-prettier": "^3.6.0",
     "eslint-plugin-babel": "^4.1.2",
@@ -159,12 +159,11 @@
     "qtip2": "2.2.0",
     "radium": "^0.25.2",
     "raw-loader": "^0.5.1",
-    "react": "~15.4.0",
-    "react-addons-test-utils": "~15.4.0",
+    "react": "~15.6",
     "react-bootstrap": "^0.32.4",
     "react-color": "^2.17.3",
     "react-datepicker": "1.6.0",
-    "react-dom": "~15.4.0",
+    "react-dom": "~15.6",
     "react-google-charts": "2",
     "react-idle-timer": "^4.2.7",
     "react-inspector": "2.3.1",
@@ -177,6 +176,7 @@
     "react-router": "^3.2.3",
     "react-select": "^1.2.1",
     "react-sticky": "^6.0.3",
+    "react-test-renderer": "~15.6",
     "react-tether": "^1.0.4",
     "react-tooltip": "^3.2.7",
     "react-virtualized": "^9.18.5",
