@@ -5,6 +5,7 @@ import color from '../../../util/color';
 import BonusLevels from './BonusLevels';
 import ProjectWidgetWithData from '@cdo/apps/templates/projects/ProjectWidgetWithData';
 import {stageOfBonusLevels} from './shapes';
+import StageExtrasNotification from './StageExtrasNotification';
 
 const styles = {
   h2: {
@@ -17,30 +18,36 @@ const styles = {
 export default class StageExtras extends React.Component {
   static propTypes = {
     stageNumber: PropTypes.number.isRequired,
+    nextStageNumber: PropTypes.number,
     nextLevelPath: PropTypes.string.isRequired,
     showProjectWidget: PropTypes.bool,
     projectTypes: PropTypes.arrayOf(PropTypes.string),
     bonusLevels: PropTypes.arrayOf(PropTypes.shape(stageOfBonusLevels)),
     sectionId: PropTypes.number,
-    userId: PropTypes.number
+    userId: PropTypes.number,
+    showStageExtrasWarning: PropTypes.bool
   };
 
   render() {
     const {
       stageNumber,
+      nextStageNumber,
       nextLevelPath,
       bonusLevels,
       sectionId,
       userId,
       showProjectWidget,
-      projectTypes
+      projectTypes,
+      showStageExtrasWarning
     } = this.props;
     const nextMessage = /stage/.test(nextLevelPath)
-      ? msg.extrasNextLesson({number: stageNumber + 1})
+      ? msg.extrasNextLesson({number: nextStageNumber})
       : msg.extrasNextFinish();
 
     return (
       <div>
+        {showStageExtrasWarning && <StageExtrasNotification />}
+
         <h1>{msg.extrasStageNumberCompleted({number: stageNumber})}</h1>
 
         <h2 style={styles.h2}>{msg.continue()}</h2>
