@@ -40,10 +40,11 @@ export default function initPage(scriptEditorData) {
         }))
     }));
   const locales = scriptEditorData.locales;
+  const flexCategoryMap = scriptEditorData.flex_category_map;
 
   registerReducers({...reducers, isRtl});
   const store = getStore();
-  store.dispatch(init(stages, scriptEditorData.levelKeyList));
+  store.dispatch(init(stages, scriptEditorData.levelKeyList, flexCategoryMap));
 
   const teacherResources = (scriptData.teacher_resources || []).map(
     ([type, link]) => ({type, link})
@@ -84,6 +85,7 @@ export default function initPage(scriptEditorData) {
         versionYear={scriptData.version_year}
         scriptFamilies={scriptEditorData.script_families}
         versionYearOptions={scriptEditorData.version_year_options}
+        isLevelbuilder={scriptEditorData.is_levelbuilder}
       />
     </Provider>,
     document.querySelector('.edit_container')
