@@ -2,6 +2,7 @@ import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 import React from 'react';
 import FontAwesome from '../../templates/FontAwesome';
+import FirebaseStorage from '../firebaseStorage';
 import msg from '@cdo/locale';
 import {showPreview} from '../redux/data';
 import PreviewModal from './PreviewModal';
@@ -16,6 +17,7 @@ class LibraryTable extends React.Component {
   static propTypes = {
     name: PropTypes.string.isRequired,
     description: PropTypes.string.isRequired,
+    current: PropTypes.bool.isRequired,
 
     // from redux dispatch
     onShowPreview: PropTypes.func.isRequired
@@ -29,6 +31,22 @@ class LibraryTable extends React.Component {
     this.setState({
       collapsed: !this.state.collapsed
     });
+
+  importTable = () => {
+    if (this.props.current) {
+      // TODO: Implement current tables (see STAR-615)
+    } else {
+      FirebaseStorage.copyStaticTable(
+        this.props.name,
+        () => {
+          console.log('success');
+        },
+        err => {
+          console.log(err);
+        }
+      );
+    }
+  };
 
   render() {
     const icon = this.state.collapsed ? 'caret-right' : 'caret-down';
@@ -47,7 +65,9 @@ class LibraryTable extends React.Component {
             >
               {msg.preview()}
             </button>
-            <button type="button"> {msg.import()} </button>
+            <button type="button" onClick={this.importTable}>
+              {msg.import()}
+            </button>
           </div>
         )}
         <PreviewModal tableName={this.props.name} />
