@@ -7,7 +7,7 @@ import i18n from '@cdo/locale';
  * A search box that loads a Google Location Search control.
  *
  * Note: Google location search requires the following line to be present in the haml where this component is used:
- * %script{type: "text/javascript", src: "https://maps.googleapis.com/maps/api/js?client=#{CDO.google_maps_client_id}&sensor=true&libraries=places,geometry&v=3.7"}
+ * %script{type: "text/javascript", src: "https://maps.googleapis.com/maps/api/js?client=#{CDO.google_maps_client_id}&sensor=true&libraries=places,geometry&v=3.37"}
  */
 export default class GoogleSchoolLocationSearchField extends React.Component {
   static propTypes = {
