@@ -246,7 +246,7 @@ class TopInstructions extends Component {
     if (serverLevelId) {
       promises.push(
         $.ajax({
-          url: `/levels/${serverLevelId}/get_rubric/`,
+          url: `/levels/${serverLevelId}/get_rubric`,
           method: 'GET',
           contentType: 'application/json;charset=UTF-8'
         }).done(data => {
