@@ -14,6 +14,10 @@ import shapes from './shapes';
 import ProtectedStatefulDiv from '../ProtectedStatefulDiv';
 import i18n from '@cdo/locale';
 import CensusTeacherBanner from '../census2017/CensusTeacherBanner';
+import DonorTeacherBanner, {
+  donorTeacherBannerOptionsShape
+} from '@cdo/apps/templates/DonorTeacherBanner';
+import experiments from '@cdo/apps/util/experiments';
 
 const styles = {
   clear: {
@@ -32,9 +36,11 @@ export default class TeacherHomepage extends Component {
     queryStringOpen: PropTypes.string,
     canViewAdvancedTools: PropTypes.bool,
     isEnglish: PropTypes.bool.isRequired,
+    ncesSchoolId: PropTypes.string,
     locale: PropTypes.string,
     showCensusBanner: PropTypes.bool.isRequired,
-    ncesSchoolId: PropTypes.string,
+    donorBannerName: PropTypes.string,
+    donorTeacherBannerOptions: donorTeacherBannerOptionsShape,
     censusQuestion: PropTypes.oneOf(['how_many_10_hours', 'how_many_20_hours']),
     teacherName: PropTypes.string,
     teacherId: PropTypes.number,
@@ -43,7 +49,8 @@ export default class TeacherHomepage extends Component {
   };
 
   state = {
-    showCensusBanner: this.props.showCensusBanner
+    showCensusBanner: this.props.showCensusBanner,
+    donorBannerName: this.props.donorBannerName
   };
 
   bindCensusBanner = banner => {
@@ -135,6 +142,28 @@ export default class TeacherHomepage extends Component {
     this.hideCensusBanner();
   };
 
+  dismissDonorTeacherBannerWithCallbacks(onSuccess, onFailure) {
+    $.ajax({
+      url: `/api/v1/users/${this.props.teacherId}/dismiss_donor_teacher_banner`,
+      type: 'post'
+    })
+      .done(onSuccess)
+      .fail(onFailure);
+  }
+
+  logDismissDonorTeacherBannerError = xhr => {
+    console.error(
+      `Failed to dismiss donor teacher banner! ${xhr.responseText}`
+    );
+  };
+
+  dismissDonorTeacherBanner() {
+    this.dismissDonorTeacherBannerWithCallbacks(
+      null,
+      this.logDismissDonorTeacherBannerError
+    );
+  }
+
   componentDidMount() {
     // The component used here is implemented in legacy HAML/CSS rather than React.
     $('#teacher_reminders')
@@ -155,6 +184,7 @@ export default class TeacherHomepage extends Component {
       ncesSchoolId,
       censusQuestion,
       schoolYear,
+      donorTeacherBannerOptions,
       teacherId,
       teacherName,
       teacherEmail,
@@ -226,6 +256,16 @@ export default class TeacherHomepage extends Component {
             <br />
           </div>
         )}
+        {experiments.isEnabled('donorTeacherBanner') &&
+          isEnglish &&
+          this.state.donorBannerName && (
+            <div>
+              <DonorTeacherBanner
+                options={donorTeacherBannerOptions}
+                onDismiss={() => this.dismissDonorTeacherBanner()}
+              />
+            </div>
+          )}
         <TeacherSections queryStringOpen={queryStringOpen} locale={locale} />
         <RecentCourses
           courses={courses}
