@@ -30,6 +30,16 @@ import {Position} from '../constants';
 
 const RANDOM_VALUE = 'RAND';
 
+const TURN_BY_DIRECTION_VALUES = [
+  [msg.turnRight(), 'turnRight'],
+  [msg.turnLeft(), 'turnLeft']
+];
+
+const TURN_DIRECTION_VALUES = [
+  [msg.right(), 'turnRight'],
+  [msg.left(), 'turnLeft']
+];
+
 const POSITION_VALUES = [
   [commonMsg.positionRandom(), RANDOM_VALUE],
   [commonMsg.positionTopLeft(), Position.TOPLEFT.toString()],
@@ -41,6 +51,26 @@ const POSITION_VALUES = [
   [commonMsg.positionBottomLeft(), Position.BOTTOMLEFT.toString()],
   [commonMsg.positionBottomCenter(), Position.BOTTOMCENTER.toString()],
   [commonMsg.positionBottomRight(), Position.BOTTOMRIGHT.toString()]
+];
+
+const JUMP_BY_DIRECTION_VALUES = [
+  [msg.jumpForward(), 'jumpForward'],
+  [msg.jumpBackward(), 'jumpBackward']
+];
+
+const JUMP_DIRECTION_VALUES = [
+  [msg.forward(), 'jumpForward'],
+  [msg.backward(), 'jumpBackward']
+];
+
+const MOVE_BY_DIRECTION_VALUES = [
+  [msg.moveForward(), 'moveForward'],
+  [msg.moveBackward(), 'moveBackward']
+];
+
+const DIRECTION_VALUES = [
+  [msg.forward(), 'moveForward'],
+  [msg.backward(), 'moveBackward']
 ];
 
 // Install extensions to Blockly's language and JavaScript generator.
@@ -98,7 +128,7 @@ exports.install = function(blockly, blockInstallOptions) {
     init: function() {
       this.setHSV(184, 1.0, 0.74);
       this.appendDummyInput().appendTitle(
-        new blockly.FieldDropdown(blockly.Blocks.draw_move.DIRECTIONS),
+        new blockly.FieldDropdown(MOVE_BY_DIRECTION_VALUES),
         'DIR'
       );
       this.appendDummyInput()
@@ -123,7 +153,7 @@ exports.install = function(blockly, blockInstallOptions) {
     init: function() {
       this.setHSV(184, 1.0, 0.74);
       this.appendDummyInput().appendTitle(
-        new blockly.FieldDropdown(blockly.Blocks.draw_move.DIRECTIONS),
+        new blockly.FieldDropdown(MOVE_BY_DIRECTION_VALUES),
         'DIR'
       );
       this.appendDummyInput()
@@ -158,7 +188,7 @@ exports.install = function(blockly, blockInstallOptions) {
     init: function() {
       this.setHSV(184, 1.0, 0.74);
       this.appendDummyInput().appendTitle(
-        new blockly.FieldDropdown(blockly.Blocks.draw_turn.DIRECTIONS),
+        new blockly.FieldDropdown(TURN_BY_DIRECTION_VALUES),
         'DIR'
       );
       this.appendDummyInput()
@@ -211,7 +241,7 @@ exports.install = function(blockly, blockInstallOptions) {
     init: function() {
       this.setHSV(184, 1.0, 0.74);
       this.appendDummyInput().appendTitle(
-        new blockly.FieldDropdown(blockly.Blocks.draw_turn.DIRECTIONS),
+        new blockly.FieldDropdown(TURN_BY_DIRECTION_VALUES),
         'DIR'
       );
       this.appendDummyInput()
@@ -235,7 +265,7 @@ exports.install = function(blockly, blockInstallOptions) {
     init: function() {
       this.setHSV(184, 1.0, 0.74);
       this.appendDummyInput().appendTitle(
-        new blockly.FieldDropdown(blockly.Blocks.draw_turn.DIRECTIONS),
+        new blockly.FieldDropdown(TURN_BY_DIRECTION_VALUES),
         'DIR'
       );
       this.appendDummyInput()
@@ -289,7 +319,7 @@ exports.install = function(blockly, blockInstallOptions) {
     init: function() {
       this.setHSV(184, 1.0, 0.74);
       this.appendDummyInput().appendTitle(
-        new blockly.FieldDropdown(blockly.Blocks.draw_turn.DIRECTIONS),
+        new blockly.FieldDropdown(TURN_BY_DIRECTION_VALUES),
         'DIR'
       );
       this.appendDummyInput()
@@ -342,7 +372,7 @@ exports.install = function(blockly, blockInstallOptions) {
     init: function() {
       this.setHSV(184, 1.0, 0.74);
       this.appendDummyInput().appendTitle(
-        new blockly.FieldDropdown(blockly.Blocks.draw_turn.DIRECTIONS),
+        new blockly.FieldDropdown(TURN_BY_DIRECTION_VALUES),
         'DIR'
       );
       this.appendDummyInput()
@@ -682,24 +712,23 @@ exports.install = function(blockly, blockInstallOptions) {
     helpUrl: '',
     init: function() {
       this.setHSV(184, 1.0, 0.74);
-      this.appendValueInput('VALUE')
-        .setCheck(blockly.BlockValueType.NUMBER)
-        .appendTitle(
-          new blockly.FieldDropdown(blockly.Blocks.draw_move.DIRECTIONS),
-          'DIR'
-        );
-      this.appendDummyInput().appendTitle(msg.dots());
+      this.interpolateMsg(
+        msg.moveDirectionByPixels(),
+        () => {
+          this.appendDummyInput().appendTitle(
+            new blockly.FieldDropdown(DIRECTION_VALUES),
+            'DIR'
+          );
+        },
+        ['VALUE', 'Number', blockly.ALIGN_RIGHT],
+        blockly.ALIGN_RIGHT
+      );
       this.setInputsInline(true);
       this.setPreviousStatement(true);
       this.setNextStatement(true);
       this.setTooltip(msg.moveTooltip());
     }
   };
-
-  blockly.Blocks.draw_move.DIRECTIONS = [
-    [msg.moveForward(), 'moveForward'],
-    [msg.moveBackward(), 'moveBackward']
-  ];
 
   generator.draw_move = function() {
     // Generate JavaScript for moving forward or backwards.
@@ -721,13 +750,17 @@ exports.install = function(blockly, blockInstallOptions) {
     helpUrl: '',
     init: function() {
       this.setHSV(184, 1.0, 0.74);
-      this.appendValueInput('VALUE')
-        .setCheck(blockly.BlockValueType.NUMBER)
-        .appendTitle(
-          new blockly.FieldDropdown(blockly.Blocks.jump.DIRECTIONS),
-          'DIR'
-        );
-      this.appendDummyInput().appendTitle(msg.dots());
+      this.interpolateMsg(
+        msg.jumpByDirection(),
+        () => {
+          this.appendDummyInput().appendTitle(
+            new blockly.FieldDropdown(JUMP_DIRECTION_VALUES),
+            'DIR'
+          );
+        },
+        ['VALUE', 'Number', blockly.ALIGN_RIGHT],
+        blockly.ALIGN_RIGHT
+      );
       this.setInputsInline(true);
       this.setPreviousStatement(true);
       this.setNextStatement(true);
@@ -1000,11 +1033,6 @@ exports.install = function(blockly, blockInstallOptions) {
 
   SimpleMove.generateBlocksForAllDirections();
 
-  blockly.Blocks.jump.DIRECTIONS = [
-    [msg.jumpForward(), 'jumpForward'],
-    [msg.jumpBackward(), 'jumpBackward']
-  ];
-
   generator.jump = function() {
     // Generate JavaScript for jumping forward or backwards.
     var value =
@@ -1027,7 +1055,7 @@ exports.install = function(blockly, blockInstallOptions) {
     init: function() {
       this.setHSV(184, 1.0, 0.74);
       this.appendDummyInput().appendTitle(
-        new blockly.FieldDropdown(blockly.Blocks.jump.DIRECTIONS),
+        new blockly.FieldDropdown(JUMP_BY_DIRECTION_VALUES),
         'DIR'
       );
       this.appendDummyInput()
@@ -1053,7 +1081,7 @@ exports.install = function(blockly, blockInstallOptions) {
     init: function() {
       this.setHSV(184, 1.0, 0.74);
       this.appendDummyInput().appendTitle(
-        new blockly.FieldDropdown(blockly.Blocks.jump.DIRECTIONS),
+        new blockly.FieldDropdown(JUMP_BY_DIRECTION_VALUES),
         'DIR'
       );
       this.appendDummyInput()
@@ -1156,28 +1184,31 @@ exports.install = function(blockly, blockInstallOptions) {
     helpUrl: '',
     init: function() {
       this.setHSV(184, 1.0, 0.74);
-      this.appendValueInput('VALUE')
-        .setCheck(blockly.BlockValueType.NUMBER)
-        .addFieldHelper(blockly.BlockFieldHelper.ANGLE_HELPER, {
-          block: this,
-          directionTitle: 'DIR'
-        })
-        .appendTitle(
-          new blockly.FieldDropdown(blockly.Blocks.draw_turn.DIRECTIONS),
-          'DIR'
-        );
-      this.appendDummyInput().appendTitle(msg.degrees());
+      this.interpolateMsg(
+        msg.turnDirection(),
+        () => {
+          this.appendDummyInput().appendTitle(
+            new blockly.FieldDropdown(TURN_DIRECTION_VALUES),
+            'DIR'
+          );
+        },
+        () => {
+          this.appendValueInput('VALUE').addFieldHelper(
+            blockly.BlockFieldHelper.ANGLE_HELPER,
+            {
+              block: this,
+              directionTitle: 'DIR'
+            }
+          );
+        },
+        blockly.ALIGN_RIGHT
+      );
       this.setInputsInline(true);
       this.setPreviousStatement(true);
       this.setNextStatement(true);
       this.setTooltip(msg.turnTooltip());
     }
   };
-
-  blockly.Blocks.draw_turn.DIRECTIONS = [
-    [msg.turnRight(), 'turnRight'],
-    [msg.turnLeft(), 'turnLeft']
-  ];
 
   generator.draw_turn = function() {
     // Generate JavaScript for turning left or right.
