@@ -4,6 +4,6 @@ maintainer_email 'dev@code.org'
 license          'All rights reserved'
 description      'Installs/Configures cdo-varnish'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.3.218'
+version          '0.3.225'
 
 depends 'apt'
