@@ -48,6 +48,17 @@ class Ability
       Pd::InternationalOptIn,
       :maker_discount
     ]
+    cannot :index, Level
+
+    # If you can see a level, you can also do these things:
+    can [:embed_level, :get_rubric], Level do |level|
+      can? :read, level
+    end
+
+    # If you can update a level, you can also do these things:
+    can [:edit_blocks, :update_blocks, :update_properties], Level do |level|
+      can? :update, level
+    end
 
     if user.persisted?
       can :manage, user
@@ -221,7 +232,7 @@ class Ability
       ]
 
       # Only custom levels are editable.
-      cannot [:update, :destroy], Level do |level|
+      cannot [:clone, :update, :destroy], Level do |level|
         !level.custom?
       end
 
@@ -233,6 +244,8 @@ class Ability
     if user.persisted?
       editor_experiment = Experiment.get_editor_experiment(user)
       if editor_experiment
+        can :index, Level
+        can :clone, Level, &:custom?
         can :manage, Level, editor_experiment: editor_experiment
         can [:edit, :update], Script, editor_experiment: editor_experiment
       end
