@@ -354,7 +354,7 @@ class HomeControllerTest < ActionController::TestCase
   # TODO: remove this test when workshop_organizer is deprecated
   test 'workshop organizers see dashboard links' do
     sign_in create(:workshop_organizer, :with_terms_of_service)
-    query_count = 14
+    query_count = 15
     assert_queries query_count do
       get :home
     end
@@ -363,7 +363,7 @@ class HomeControllerTest < ActionController::TestCase
 
   test 'program managers see dashboard links' do
     sign_in create(:program_manager, :with_terms_of_service)
-    query_count = 16
+    query_count = 17
     assert_queries query_count do
       get :home
     end
@@ -372,7 +372,7 @@ class HomeControllerTest < ActionController::TestCase
 
   test 'workshop admins see dashboard links' do
     sign_in create(:workshop_admin, :with_terms_of_service)
-    query_count = 12
+    query_count = 13
     assert_queries query_count do
       get :home
     end
@@ -382,7 +382,7 @@ class HomeControllerTest < ActionController::TestCase
   test 'facilitators see dashboard links' do
     facilitator = create(:facilitator, :with_terms_of_service)
     sign_in facilitator
-    query_count = 13
+    query_count = 14
     assert_queries query_count do
       get :home
     end
@@ -391,7 +391,7 @@ class HomeControllerTest < ActionController::TestCase
 
   test 'teachers cannot see dashboard links' do
     sign_in create(:terms_of_service_teacher)
-    query_count = 12
+    query_count = 13
     assert_queries query_count do
       get :home
     end
@@ -400,7 +400,7 @@ class HomeControllerTest < ActionController::TestCase
 
   test 'workshop admins see application dashboard links' do
     sign_in create(:workshop_admin, :with_terms_of_service)
-    query_count = 12
+    query_count = 13
     assert_queries query_count do
       get :home
     end
@@ -411,7 +411,7 @@ class HomeControllerTest < ActionController::TestCase
   # TODO: remove this test when workshop_organizer is deprecated
   test 'workshop organizers who are regional partner program managers see application dashboard links' do
     sign_in create(:workshop_organizer, :as_regional_partner_program_manager, :with_terms_of_service)
-    query_count = 16
+    query_count = 17
     assert_queries query_count do
       get :home
     end
@@ -421,7 +421,7 @@ class HomeControllerTest < ActionController::TestCase
 
   test 'program managers see application dashboard links' do
     sign_in create(:program_manager, :with_terms_of_service)
-    query_count = 16
+    query_count = 17
     assert_queries query_count do
       get :home
     end
@@ -432,7 +432,7 @@ class HomeControllerTest < ActionController::TestCase
   # TODO: remove this test when workshop_organizer is deprecated
   test 'workshop organizers who are not regional partner program managers do not see application dashboard links' do
     sign_in create(:workshop_organizer, :with_terms_of_service)
-    query_count = 14
+    query_count = 15
     assert_queries query_count do
       get :home
     end
